♻️ refactor: change folder structure, edit login page markup, update svg using svgr
