+++ src/pages/Login/ui/Page.tsx
@@ -1,10 +1,13 @@
-import { Button, GoogleSocialIcon, Input, MailIcon, Logo } from "~/shared/ui";
+import { IconMail01 } from "~/shared/assets/icons";
+import { Button } from "~/shared/ui/button";
+import { Input } from "~/shared/ui/input";
+import { Logo } from "~/shared/ui/logo";
 
 export const Page = () => {
   return (
-    <main className="relative flex h-screen flex-col items-center gap-8 px-4 pb-12 lg:h-screen lg:w-screen lg:flex-row lg:gap-0 lg:p-0">
-      <div className="absolute inset-x-0 top-0 h-16 bg-[url('/images/geometric-shapes-mobile.svg')] lg:static lg:order-2 lg:block lg:aspect-[4/5] lg:h-screen lg:bg-[url('/images/geometric-shapes-desktop.svg')] lg:bg-cover" />
-      <div className="flex flex-col items-center justify-between gap-8 self-stretch pt-24 lg:flex-[1_0_0] lg:pt-0">
+    <main className="relative flex h-screen flex-col items-center gap-8 pb-12 lg:h-screen lg:w-screen lg:flex-row lg:justify-between lg:gap-0 lg:p-0">
+      <div className="g:aspect-[4/5] -order-1 h-16 w-full bg-[url('/src/pages/Login/ui/geometric-shapes-mobile.svg')] bg-center bg-no-repeat lg:order-1 lg:h-full lg:w-full lg:max-w-[756px] lg:bg-[url('/src/pages/Login/ui/geometric-shapes-desktop.svg')]" />
+      <div className="flex flex-col items-center justify-between gap-8 self-stretch px-4 pt-24 lg:min-w-[560px] lg:flex-grow lg:pt-0">
         <header className="hidden flex-col items-start gap-6 self-stretch lg:flex lg:h-24 lg:p-8">
           <Logo />
         </header>
@@ -29,7 +32,7 @@ export const Page = () => {
               <Button type="submit" size={"lg"} className="self-stretch">
                 Get started
               </Button>
-              <Button
+              {/* <Button
                 type="button"
                 variant={"secondary-gray"}
                 size={"lg"}
@@ -37,16 +40,16 @@ export const Page = () => {
                 className="self-stretch"
               >
                 Sign up with Google
-              </Button>
+              </Button> */}
             </div>
           </form>
         </section>
         <footer className="hidden lg:flex lg:h-24 lg:items-end lg:justify-between lg:self-stretch lg:p-9">
           <p className="text-sm text-gray-600">© Brello 2023</p>
-          <div className="flex items-center gap-2 text-gray-600">
-            <MailIcon />
-            <p className="text-sm">help@brello.io</p>
-          </div>
+          <p className="flex items-center gap-2 text-sm text-gray-600">
+            <IconMail01 className="h-4 w-4 stroke-gray-600" />
+            help@brello.io
+          </p>
         </footer>
       </div>
     </main>
