+++ src/shared/assets/icons/index.ts
@@ -1,0 +1,4 @@
+import IconMail01 from "./mail-01.svg?react";
+import IconGoogleSocial from "./google-social-icon.svg?react";
+
+export { IconMail01, IconGoogleSocial };
