+++ src/shared/assets/images/index.ts
@@ -1,0 +1,3 @@
+import Logotype from "./logotype.svg?react";
+
+export { Logotype };
