--- src/shared/ui/icons/MailIcon.tsx
@@ -1,19 +1,0 @@
-export const MailIcon = ({ size = 16 }: { size?: number }) => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width={size}
-      height={size}
-      viewBox={`0 0 16 16`}
-      fill="none"
-    >
-      <path
-        d="M1.33331 4.66666L6.77659 8.47695C7.21737 8.7855 7.43776 8.93977 7.67749 8.99953C7.88924 9.05231 8.11072 9.05231 8.32247 8.99953C8.5622 8.93977 8.78259 8.7855 9.22337 8.47695L14.6666 4.66666M4.53331 13.3333H11.4666C12.5868 13.3333 13.1468 13.3333 13.5746 13.1153C13.951 12.9236 14.2569 12.6176 14.4487 12.2413C14.6666 11.8135 14.6666 11.2534 14.6666 10.1333V5.86666C14.6666 4.74655 14.6666 4.1865 14.4487 3.75868C14.2569 3.38235 13.951 3.07639 13.5746 2.88464C13.1468 2.66666 12.5868 2.66666 11.4666 2.66666H4.53331C3.41321 2.66666 2.85316 2.66666 2.42533 2.88464C2.04901 3.07639 1.74305 3.38235 1.5513 3.75868C1.33331 4.1865 1.33331 4.74655 1.33331 5.86666V10.1333C1.33331 11.2534 1.33331 11.8135 1.5513 12.2413C1.74305 12.6176 2.04901 12.9236 2.42533 13.1153C2.85316 13.3333 3.41321 13.3333 4.53331 13.3333Z"
-        stroke="currentColor"
-        strokeWidth="1.33333"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-};
