--- src/shared/ui/icons/index.ts
@@ -1,2 +1,0 @@
-export { GoogleSocialIcon } from "./GoogleSocialIcon";
-export { MailIcon } from "./MailIcon";
