--- src/shared/ui/svgs/Logotype.tsx
@@ -1,30 +1,0 @@
-export const Logotype = () => {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="101"
-      height="32"
-      viewBox="0 0 101 32"
-      fill="none"
-    >
-      <path
-        d="M0.517029 24V6.54544H7.50567C8.78976 6.54544 9.86078 6.73578 10.7187 7.11646C11.5767 7.49715 12.2216 8.02555 12.6534 8.70169C13.0852 9.37215 13.3011 10.1449 13.3011 11.0199C13.3011 11.7017 13.1648 12.3011 12.892 12.8182C12.6193 13.3295 12.2443 13.75 11.767 14.0795C11.2954 14.4034 10.7557 14.6335 10.1477 14.7699V14.9403C10.8125 14.9687 11.4346 15.1562 12.0142 15.5028C12.5994 15.8494 13.0738 16.3352 13.4375 16.9602C13.8011 17.5795 13.9829 18.3182 13.9829 19.1761C13.9829 20.1023 13.7528 20.929 13.2926 21.6562C12.8381 22.3778 12.1648 22.9489 11.2727 23.3693C10.3807 23.7898 9.28123 24 7.97442 24H0.517029ZM4.20737 20.9829H7.21589C8.2443 20.9829 8.9943 20.7869 9.46589 20.3949C9.93748 19.9971 10.1733 19.4687 10.1733 18.8096C10.1733 18.3267 10.0568 17.9006 9.82385 17.5312C9.59089 17.1619 9.25851 16.8721 8.82669 16.6619C8.40055 16.4517 7.89203 16.3466 7.30112 16.3466H4.20737V20.9829ZM4.20737 13.8494H6.94317C7.44885 13.8494 7.89771 13.7614 8.28976 13.5852C8.68748 13.4034 8.99998 13.1477 9.22726 12.8182C9.46021 12.4886 9.57669 12.0937 9.57669 11.6335C9.57669 11.0028 9.35226 10.4943 8.90339 10.1079C8.46021 9.72158 7.82953 9.52839 7.01135 9.52839H4.20737V13.8494Z"
-        fill="black"
-      />
-      <path
-        d="M16.316 24V10.9091H19.8359V13.1932H19.9723C20.2109 12.3807 20.6115 11.767 21.174 11.3523C21.7365 10.9318 22.3842 10.7216 23.1172 10.7216C23.299 10.7216 23.495 10.7329 23.7052 10.7557C23.9155 10.7784 24.1001 10.8096 24.2592 10.8494V14.071C24.0888 14.0199 23.853 13.9744 23.5518 13.9346C23.2507 13.8949 22.9751 13.875 22.7251 13.875C22.191 13.875 21.7138 13.9915 21.2933 14.2244C20.8785 14.4517 20.549 14.7699 20.3047 15.179C20.066 15.5881 19.9467 16.0596 19.9467 16.5937V24H16.316Z"
-        fill="black"
-      />
-      <path
-        d="M31.4929 24.2557C30.1463 24.2557 28.9872 23.9829 28.0156 23.4375C27.0497 22.8864 26.3054 22.1079 25.7827 21.1023C25.2599 20.0909 24.9986 18.8949 24.9986 17.5142C24.9986 16.1676 25.2599 14.9858 25.7827 13.9687C26.3054 12.9517 27.0412 12.1591 27.99 11.5909C28.9446 11.0227 30.0639 10.7386 31.348 10.7386C32.2116 10.7386 33.0156 10.8778 33.7599 11.1562C34.5099 11.429 35.1633 11.8409 35.7202 12.392C36.2827 12.9432 36.7202 13.6363 37.0327 14.4716C37.3452 15.3011 37.5014 16.2727 37.5014 17.3864V18.3835H26.4474V16.1335H34.0838C34.0838 15.6108 33.9702 15.1477 33.7429 14.7443C33.5156 14.3409 33.2003 14.0256 32.7969 13.7983C32.3991 13.5653 31.9361 13.4489 31.4077 13.4489C30.8565 13.4489 30.3679 13.5767 29.9417 13.8324C29.5213 14.0824 29.1917 14.4204 28.9531 14.8466C28.7145 15.267 28.5923 15.7358 28.5866 16.2528V18.392C28.5866 19.0398 28.7059 19.5994 28.9446 20.071C29.1889 20.5426 29.5327 20.9062 29.9758 21.1619C30.419 21.4176 30.9446 21.5454 31.5525 21.5454C31.9559 21.5454 32.3253 21.4886 32.6605 21.375C32.9957 21.2614 33.2827 21.0909 33.5213 20.8636C33.7599 20.6363 33.9417 20.3579 34.0667 20.0284L37.4247 20.25C37.2542 21.0568 36.9048 21.7613 36.3764 22.3636C35.8537 22.9602 35.1775 23.4261 34.348 23.7614C33.5241 24.0909 32.5724 24.2557 31.4929 24.2557Z"
-        fill="black"
-      />
-      <path d="M43.5014 6.54544V24H39.8707V6.54544H43.5014Z" fill="black" />
-      <path d="M50.0405 6.54544V24H46.4098V6.54544H50.0405Z" fill="black" />
-      <path
-        d="M58.8636 24.2557C57.5398 24.2557 56.3949 23.9744 55.429 23.4119C54.4687 22.8437 53.7273 22.054 53.2045 21.0426C52.6818 20.0256 52.4204 18.8466 52.4204 17.5057C52.4204 16.1534 52.6818 14.9716 53.2045 13.9602C53.7273 12.9432 54.4687 12.1534 55.429 11.5909C56.3949 11.0227 57.5398 10.7386 58.8636 10.7386C60.1875 10.7386 61.3295 11.0227 62.2898 11.5909C63.2557 12.1534 64 12.9432 64.5227 13.9602C65.0454 14.9716 65.3068 16.1534 65.3068 17.5057C65.3068 18.8466 65.0454 20.0256 64.5227 21.0426C64 22.054 63.2557 22.8437 62.2898 23.4119C61.3295 23.9744 60.1875 24.2557 58.8636 24.2557ZM58.8807 21.4432C59.4829 21.4432 59.9858 21.2727 60.3892 20.9318C60.7926 20.5852 61.0966 20.1136 61.3011 19.517C61.5113 18.9204 61.6165 18.2415 61.6165 17.4801C61.6165 16.7187 61.5113 16.0398 61.3011 15.4432C61.0966 14.8466 60.7926 14.375 60.3892 14.0284C59.9858 13.6818 59.4829 13.5085 58.8807 13.5085C58.2727 13.5085 57.7613 13.6818 57.3466 14.0284C56.9375 14.375 56.6278 14.8466 56.4176 15.4432C56.2131 16.0398 56.1108 16.7187 56.1108 17.4801C56.1108 18.2415 56.2131 18.9204 56.4176 19.517C56.6278 20.1136 56.9375 20.5852 57.3466 20.9318C57.7613 21.2727 58.2727 21.4432 58.8807 21.4432Z"
-        fill="black"
-      />
-    </svg>
-  );
-};
