--- src/shared/ui/svgs/index.ts
@@ -1,3 +1,0 @@
-export { GeometricShapesDesktop } from "./GeometricShapesDesktop";
-export { GeometricShapesMobile } from "./GeometricShapesMobile";
-export { Logotype } from "./Logotype";
